Implement `removeSavedConnection`.
Since the backend checks for an exact match in `.saved`, and the saved connection on the frontend may include metadata that isn't in the file, it's necessary to strip the metadata before sending the connection. Might as well do this for `add` and `select` as well to be consistent, even though it isn't technically necessary (I only know that because I know the implementation details).